#111 feat: filter params 커스텀 훅으로 분리
